cast percent as int

+++ cmd/count.go
@@ -180,7 +180,7 @@ func runCount(cmd *cobra.Command, args []string) error {
 
 		// Calculate and add the percentage for each language.
 		for _, langData := range sortedLanguages {
-			percentage := float64(langData.Count) / float64(totalRepos) * 100
+			percentage := int(float64(langData.Count) / float64(totalRepos) * 100)
 			rows = append(rows, []string{langData.Language, fmt.Sprintf("%d", langData.Count), fmt.Sprintf("%d%%", percentage)})
 		}
